Update Tasks test suite leafs and Tasks Full Test Suite

+++ test/api/test-suite/tasks_cronjob.py
@@ -34,6 +34,10 @@ def get():
     print 'Get tasks-cronjob --> Failed!'
 
 def post():
+  r = requests.get(url, auth = auth)
+  result = json.loads(r.text)
+  if len(result) > 0:
+    delete()
   r = requests.post(url, auth = auth, data = json.dumps(payload), headers = headers)
   if r.status_code == 201:
     result = json.loads(r.text)
@@ -41,6 +45,7 @@ def post():
     return str(result['id'])+'/'
   else:
     print 'Create tasks-cronjob --> Failed!'
+    return ''
 
 def put():
   id = post()
@@ -51,12 +56,22 @@ def put():
     print 'Update tasks-cronjob --> Failed!'
 
 def delete():
-  id = post()
-  r = requests.delete(url+id, auth = auth)
-  if r.status_code == 204:
-    print 'Delete tasks-cronjob --> Succeeded!'
-  else:
-    print 'Delete tasks-cronjob --> Failed!'
+  r = requests.get(url, auth = auth)
+  result = json.loads(r.text)
+  i = 0
+  for i in range(0,len(result)):
+    r = requests.delete(url+str(result[i]['id'])+'/', auth = auth)
+    if r.status_code == 204:
+      print 'Delete tasks-initshutdown --> Succeeded!'
+    else:
+      print 'Delete tasks-initshutdown --> Failed!'
+  if len(result) == 0:
+    id = post()
+    r = requests.delete(url+id, auth = auth)
+    if r.status_code == 204:
+      print 'Delete tasks-initshutdown --> Succeeded!'
+    else:
+      print 'Delete tasks-initshutdown --> Failed!'
 
 def run():
   id = post()
